Create generate candidate with its word-index
Use Stanford to candidates first (keep the order as the same as within the sentence, also keep repeat candidate), then scan the sentence words from beginning to the end, so that give the extracted candidates the index which are their word-level position of the sentence. StanfordNLP and swissom_ai need to be ready before using.